Revert "Updating GNMT"
This reverts commit 072523d5b21444a97b82876bf4c6d1d3e36ba255.

+++ PyTorch/Translation/GNMT/seq2seq/train/smoothing.py
@@ -1,7 +1,6 @@
 import torch
 import torch.nn as nn
 
-
 class LabelSmoothing(nn.Module):
     """
     NLL loss with label smoothing.
@@ -19,8 +18,7 @@ class LabelSmoothing(nn.Module):
         self.smoothing = smoothing
 
     def forward(self, x, target):
-        logprobs = torch.nn.functional.log_softmax(x, dim=-1,
-                                                   dtype=torch.float32)
+        logprobs = torch.nn.functional.log_softmax(x, dim=-1)
 
         non_pad_mask = (target != self.padding_idx)
         nll_loss = -logprobs.gather(dim=-1, index=target.unsqueeze(1))
